service/requirements.txt
asgiref==3.6.0
    # via django
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via pip-tools
django==4.1.5
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.4
    # via requests
packaging==23.0
    # via build
pip-tools==6.12.1
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytz==2022.7.1
    # via djangorestframework
requests==2.28.2
    # via -r requirements.in
sqlparse==0.4.3
    # via django
tomli==2.0.1
    # via build
urllib3==1.26.14
    # via requests
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
